Fix OoyalaSkinSDK Xcode project
With the update of React Native we were missing to add dependencies in the OoyalaSkinSDK xcode project.

Also set the enable modules setting to true since it is true for regular React Native apps, without it the project was still failing to build. Not sure about the details but it fixed it.